Move helpers to PathMapper.scala
